got rid of the annoying index folder naming convention... lol I'm no longer a fan of it
