+++ src/components/app.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Sites from '../home-page/sites';
-import OpenSource from '../home-page/open-source';
-import Blog from '../home-page/blog';
-import Footer from '../home-page/footer';
-import LandingScreen from '../home-page/landing-screen';
-import Contact from '../home-page/contact';
+import Sites from './home-page/sites';
+import OpenSource from './home-page/open-source';
+import Blog from './home-page/blog';
+import Footer from './home-page/footer';
+import LandingScreen from './home-page/landing-screen';
+import Contact from './home-page/contact';
 
 export default class App extends React.Component {
   componentDidMount() {
