+++ src/components/home-page/background-particles.jsx
