+++ src/components/home-page/footer.jsx
