--- src/index.css
@@ -1,4 +1,0 @@
-body {
-  margin: 0px;
-  font-family: sans-serif;
-}
